# api/openapi.json
{
  "components": {
    "schemas": {
...
      "AuditEventResponse": {
        "properties": {
          "action": {
            "type": "string"
          },
          "actorEmail": {
            "description": "PII. `[PII_REDACTED]` on a rejected unauthenticated attempt.",
            "nullable": true,
            "type": "string"
          },
          "actorId": {
            "nullable": true,
            "type": "string"
          },
          "actorRole": {
            "nullable": true,
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "hash": {
            "nullable": true,
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "ipAddress": {
            "nullable": true,
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "description": "Request body after redaction: secrets become [REDACTED], PII fields become [PII_REDACTED]. Free-form by nature — this is the one response field a generated client cannot narrow.",
            "nullable": true,
            "type": "object"
          },
          "prevHash": {
            "nullable": true,
            "type": "string"
          },
          "resourceId": {
            "nullable": true,
            "type": "string"
          },
          "resourceType": {
            "nullable": true,
            "type": "string"
          },
          "sequence": {
            "description": "Monotonic. The hash chain is ordered by this.",
            "type": "number"
          },
          "userAgent": {
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "id",
          "sequence",
          "actorId",
          "actorEmail",
          "actorRole",
          "action",
          "resourceType",
          "resourceId",
          "ipAddress",
          "userAgent",
          "prevHash",
          "hash",
          "createdAt"
        ],
        "type": "object"
      },
      "AuditListResponse": {
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/AuditEventResponse"
            },
            "type": "array"
          },
          "total": {
            "description": "Total matching the filter, not the page size.",
            "type": "number"
          }
        },
        "required": [
          "events",
          "total"
        ],
        "type": "object"
      },
...
      "ChainVerificationResponse": {
        "properties": {
          "firstBrokenSequence": {
            "description": "Present only when `valid` is false. Absent, not null, on a good chain.",
            "type": "number"
          },
          "totalEvents": {
            "type": "number"
          },
          "valid": {
            "type": "boolean"
          }
        },
        "required": [
          "valid",
          "totalEvents"
        ],
        "type": "object"
      },
      "CreateAttestationDto": {
        "properties": {
          "circulatingSupply": {
            "minimum": 1,
            "type": "number"
          },
          "method": {
            "maxLength": 120,
            "minLength": 2,
            "type": "string"
          },
          "notes": {
            "maxLength": 2000,
            "type": "string"
          },
          "periodEnd": {
            "type": "string"
          },
          "periodStart": {
            "type": "string"
          },
          "programCode": {
            "maxLength": 32,
            "minLength": 1,
            "type": "string"
          },
          "verifiedReserve": {
            "minimum": 1,
            "type": "number"
          }
        },
        "required": [
          "programCode",
          "circulatingSupply",
          "verifiedReserve",
          "method",
          "periodStart",
          "periodEnd"
        ],
        "type": "object"
      },
      "CreateKycCaseDto": {
        "properties": {
          "country": {
            "maxLength": 56,
            "minLength": 2,
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "legalName": {
            "maxLength": 120,
            "minLength": 2,
            "type": "string"
          },
          "subjectType": {
            "enum": [
              "person",
              "entity"
            ],
            "type": "string"
          }
        },
        "required": [
          "subjectType",
          "legalName",
          "country"
        ],
        "type": "object"
      },
...
      "InvestorAuthResult": {
        "properties": {
          "accessToken": {
            "description": "Investor-domain JWT (typ=investor). Not accepted by any admin route.",
            "type": "string"
          },
          "investor": {
            "$ref": "#/components/schemas/InvestorPublic"
          }
        },
        "required": [
          "accessToken",
          "investor"
        ],
        "type": "object"
      },
      "InvestorKycSummary": {
        "properties": {
          "riskLevel": {
            "enum": [
              "low",
              "medium",
              "high",
              "unrated"
            ],
            "type": "string"
          },
          "sanctions": {
            "description": "Never a real screening result. `not_screened` when nothing ran, `clear_stub` when the illustrative screener did.",
            "type": "string"
          },
          "status": {
            "description": "`not_started` is synthesised when no case exists; the other values come from the case row.",
            "enum": [
              "not_started",
              "pending",
              "in_review",
              "approved",
              "rejected"
            ],
            "type": "string"
          }
        },
        "required": [
          "status",
          "riskLevel",
          "sanctions"
        ],
        "type": "object"
      },
...
      "InvestorPublic": {
        "properties": {
          "email": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "id": {
            "description": "Null when the caller has a valid token but no investor row — `status` falls back to an email-only profile rather than 404ing.",
            "nullable": true,
            "type": "string"
          },
          "memberSince": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "fullName"
        ],
        "type": "object"
      },
      "InvestorStatus": {
        "properties": {
          "kyc": {
            "$ref": "#/components/schemas/InvestorKycSummary"
          },
          "profile": {
            "$ref": "#/components/schemas/InvestorPublic"
          },
          "programs": {
            "items": {
              "$ref": "#/components/schemas/InvestorProgramSummary"
            },
            "type": "array"
          },
          "waitlist": {
            "allOf": [
              {
                "$ref": "#/components/schemas/InvestorWaitlistSummary"
              }
            ],
            "nullable": true
          }
        },
        "required": [
          "profile",
          "waitlist",
          "kyc",
          "programs"
        ],
        "type": "object"
      },
...
      "KycCaseDetail": {
        "properties": {
          "country": {
            "description": "ISO 3166-1 alpha-2.",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "email": {
            "description": "PII. Present on this role-guarded detail route only; the case list omits it (invariant 23).",
            "nullable": true,
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "legalName": {
            "type": "string"
          },
          "notes": {
            "nullable": true,
            "type": "string"
          },
          "reviewedAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "reviewedBy": {
            "nullable": true,
            "type": "string"
          },
          "riskLevel": {
            "enum": [
              "low",
              "medium",
              "high",
              "unrated"
            ],
            "nullable": true,
            "type": "string"
          },
          "sanctions": {
            "description": "Persisted screening outcome. `clear_stub` is written by the illustrative screener and is not a real screening result. Null means never screened.",
            "nullable": true,
            "type": "string"
          },
          "status": {
            "enum": [
              "pending",
              "in_review",
              "approved",
              "rejected"
            ],
            "type": "string"
          },
          "subjectType": {
            "enum": [
              "person",
              "entity"
            ],
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "subjectType",
          "legalName",
          "country",
          "status",
          "createdAt",
          "updatedAt"
        ],
        "type": "object"
      },
      "KycCaseListItem": {
        "properties": {
          "country": {
            "description": "ISO 3166-1 alpha-2.",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "legalName": {
            "type": "string"
          },
          "notes": {
            "nullable": true,
            "type": "string"
          },
          "reviewedAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "reviewedBy": {
            "nullable": true,
            "type": "string"
          },
          "riskLevel": {
            "enum": [
              "low",
              "medium",
              "high",
              "unrated"
            ],
            "nullable": true,
            "type": "string"
          },
          "sanctions": {
            "description": "Persisted screening outcome. `clear_stub` is written by the illustrative screener and is not a real screening result. Null means never screened.",
            "nullable": true,
            "type": "string"
          },
          "status": {
            "enum": [
              "pending",
              "in_review",
              "approved",
              "rejected"
            ],
            "type": "string"
          },
          "subjectType": {
            "enum": [
              "person",
              "entity"
            ],
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "subjectType",
          "legalName",
          "country",
          "status",
          "createdAt",
          "updatedAt"
        ],
        "type": "object"
      },
      "KycScreenResult": {
        "properties": {
          "caseId": {
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "provider": {
            "description": "Always `stub`. No provider is called.",
            "enum": [
              "stub"
            ],
            "type": "string"
          },
          "result": {
            "description": "Always `clear`. Not a screening result.",
            "enum": [
              "clear"
            ],
            "type": "string"
          },
          "screenedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "caseId",
          "provider",
          "result",
          "screenedAt",
          "note"
        ],
        "type": "object"
      },
      "KycStats": {
        "properties": {
          "byStatus": {
            "items": {
              "$ref": "#/components/schemas/KycStatusCount"
            },
            "type": "array"
          },
          "total": {
            "type": "number"
          }
        },
        "required": [
          "total",
          "byStatus"
        ],
        "type": "object"
      },
...
      "LoginInvestorDto": {
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "type": "object"
      },
...
      "RegisterInvestorDto": {
        "properties": {
          "email": {
            "type": "string"
          },
          "fullName": {
            "maxLength": 120,
            "minLength": 2,
            "type": "string"
          },
          "password": {
            "minLength": 8,
            "type": "string"
          }
        },
        "required": [
          "fullName",
          "email",
          "password"
        ],
        "type": "object"
      },
...
      "ReviewKycCaseDto": {
        "properties": {
          "notes": {
            "maxLength": 2000,
            "type": "string"
          },
          "riskLevel": {
            "enum": [
              "low",
              "medium",
              "high",
              "unrated"
            ],
            "type": "string"
          },
          "status": {
            "enum": [
              "pending",
              "in_review",
              "approved",
              "rejected"
            ],
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "description": "Admin domain (typ=admin, JWT_SECRET) on admin routes; investor domain (typ=investor, INVESTOR_JWT_SECRET) on /investor/me and /investor/status. The two are disjoint and neither is accepted by the other.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "ReserveChain API.\n\nIN DEVELOPMENT. No tokens are offered or sold.\n\nProof-of-Reserves, redemption, wallet linking and token purchase are\npublished as a contract only. Every operation tagged `inactive` returns\n501 and is marked deprecated so a generated client warns at the call site.\nThey require written authorization before they do anything.\n\nNo example values are supplied anywhere in this document. Every response\nfield is described by type and constraint only.",
    "title": "ReserveChain API",
    "version": "0.1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/audit": {
      "get": {
        "operationId": "Audit_list",
        "parameters": [
          {
            "in": "query",
            "name": "skip",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "take",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "actorId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "action",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "resourceType",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "fromDate",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "toDate",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditListResponse"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "audit"
        ]
      }
    },
    "/api/audit/verify": {
      "get": {
        "operationId": "Audit_verify",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChainVerificationResponse"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "audit"
        ]
      }
    },
...
    "/api/investor/login": {
      "post": {
        "operationId": "Investor_login",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginInvestorDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestorAuthResult"
                }
              }
            },
            "description": ""
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestorAuthResult"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "investor"
        ]
      }
    },
    "/api/investor/me": {
      "get": {
        "operationId": "Investor_me",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestorPublic"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "investor"
        ]
      }
    },
    "/api/investor/register": {
      "post": {
        "operationId": "Investor_register",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterInvestorDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestorAuthResult"
                }
              }
            },
            "description": ""
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestorAuthResult"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "investor"
        ]
      }
    },
    "/api/investor/status": {
      "get": {
        "operationId": "Investor_status",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestorStatus"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "investor"
        ]
      }
    },
    "/api/kyc/cases": {
      "get": {
        "operationId": "Kyc_list",
        "parameters": [
          {
            "in": "query",
            "name": "take",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/KycCaseListItem"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "kyc"
        ]
      },
      "post": {
        "operationId": "Kyc_create",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateKycCaseDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KycCaseDetail"
                }
              }
            },
            "description": ""
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KycCaseDetail"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "kyc"
        ]
      }
    },
    "/api/kyc/cases/{id}": {
      "get": {
        "operationId": "Kyc_get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KycCaseDetail"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "kyc"
        ]
      }
    },
    "/api/kyc/cases/{id}/review": {
      "post": {
        "operationId": "Kyc_review",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewKycCaseDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KycCaseDetail"
                }
              }
            },
            "description": ""
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KycCaseDetail"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "kyc"
        ]
      }
    },
    "/api/kyc/cases/{id}/screen": {
      "post": {
        "operationId": "Kyc_screen",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KycScreenResult"
                }
              }
            },
            "description": ""
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KycScreenResult"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "kyc"
        ]
      }
    },
    "/api/kyc/stats": {
      "get": {
        "operationId": "Kyc_stats",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KycStats"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "kyc"
        ]
      }
    },
...
    "/api/por/attestations": {
      "get": {
        "deprecated": true,
        "operationId": "ProofOfReserves_list",
        "parameters": [
          {
            "in": "query",
            "name": "take",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "INACTIVE (P11). Gated by PROOF_OF_RESERVES_ENABLED; with the flag on the service still refuses. Requires written authorization before it does anything."
          }
        },
        "tags": [
          "inactive"
        ]
      },
      "post": {
        "deprecated": true,
        "operationId": "ProofOfReserves_create",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAttestationDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "501": {
            "description": "INACTIVE (P11). Gated by PROOF_OF_RESERVES_ENABLED; with the flag on the service still refuses. Requires written authorization before it does anything."
          }
        },
        "tags": [
          "inactive"
        ]
      }
    },
    "/api/por/attestations/{id}/publish": {
      "post": {
        "deprecated": true,
        "operationId": "ProofOfReserves_publish",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "INACTIVE (P11). Gated by PROOF_OF_RESERVES_ENABLED; with the flag on the service still refuses. Requires written authorization before it does anything."
          }
        },
        "tags": [
          "inactive"
        ]
      }
    },
...
  },
  "servers": [],
...
}