# api/src/audit/audit.controller.ts
import { Controller, Get, Query, UseGuards } from '@nestjs/common';
import { AuditService } from './audit.service';
import { JwtAuthGuard } from '../common/guards/jwt-auth.guard';
import { RolesGuard } from '../common/guards/roles.guard';
import { Roles } from '../common/decorators/roles.decorator';
import { Role } from '../common/enums/role.enum';
import { ApiBearerAuth, ApiOkResponse, ApiTags } from '@nestjs/swagger';
import {
  AuditListResponse,
  ChainVerificationResponse,
} from './dto/audit.response.dto';

@ApiTags('audit')
@ApiBearerAuth()
@Controller('audit')
@UseGuards(JwtAuthGuard, RolesGuard)
@Roles(Role.ADMIN, Role.COMPLIANCE)
export class AuditController {
  constructor(private readonly service: AuditService) {}

  @ApiOkResponse({ type: AuditListResponse })
  @Get()
  async list(
    @Query('skip') skip?: string,
    @Query('take') take?: string,
    @Query('actorId') actorId?: string,
    @Query('action') action?: string,
    @Query('resourceType') resourceType?: string,
    @Query('fromDate') fromDate?: string,
    @Query('toDate') toDate?: string,
  ) {
    return this.service.list({
      skip: skip ? Number(skip) : 0,
      take: take ? Number(take) : 50,
      actorId,
      action,
      resourceType,
      fromDate: fromDate ? new Date(fromDate) : undefined,
      toDate: toDate ? new Date(toDate) : undefined,
    });
  }

  @ApiOkResponse({ type: ChainVerificationResponse })
  @Get('verify')
  async verify() {
    return this.service.verifyChain();
  }
}